% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{apply_ref_method}
\alias{apply_ref_method}
\title{Apply a reference method}
\usage{
apply_ref_method(.valve_data, .l, .n)
}
\arguments{
\item{.valve_data}{\code{valve_data}}

\item{.l}{a list containing \code{rt}, \code{zf}, and \code{zfa}}

\item{.n}{a string naming the method}
}
\description{
Apply a reference method
}
