% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{changepoint}
\alias{changepoint}
\alias{id_changepoint}
\alias{pbca_minpoint}
\alias{maxpoint}
\title{Functions for detecting changepoints laser transects}
\usage{
id_changepoint(x, .var, .use_up_to_row, .method = "AMOC", .outer = FALSE)

pbca_minpoint(x, .use_up_to_row, .outer = FALSE)

maxpoint(x, .var, .use_up_to_row, .threshold = 1, .outer = FALSE)
}
\arguments{
\item{x}{a chemistry dataset}

\item{.use_up_to_row}{how many rows of data to used in the changepoint detection}

\item{.method}{passed to \code{\link[changepoint]{cpt.meanvar}}}
}
\description{
Uses \code{\link[changepoint]{cpt.meanvar}} to find the
changepoint in \code{Ca43_CPS} or \code{Pb208_CPS} or their ratio
at the epoxy/valve transition.
}
\section{Functions}{
\itemize{
\item \code{pbca_minpoint}: detect change in Pb/Ca by minimum value

\item \code{maxpoint}: detect change by maximum
}}

