% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_data_functions.R
\name{create_analysis_data_preparer}
\alias{create_analysis_data_preparer}
\title{Create a function that prepares data for analysis}
\usage{
create_analysis_data_preparer(
  prepared_analysis_data,
  analysis_grouping = rlang::syms(c("species", "element", "signal", "contrast")),
  adata_filter_FUN = filter_analysis_data,
  Z_maker_FUN = Z_maker
)
}
\arguments{
\item{prepared_analysis_data}{TODO}

\item{analysis_grouping}{a list of variables which define the final grouping of
the data returned by the inner function.}
}
\value{
a function of \code{contrast}, \code{group_by_valve},
\code{test_data_FUN}, and \code{...}. where \code{contrast} is the name
of the contrast of interest. The  \code{contrast} determines the
\code{identifiers} argument passed to \code{filter_analysis_data}.
\code{group_by_valve} is an indicator (defaulting to \code{FALSE}) for
whether to group observation by valve (TRUE) or by transect within valve
(TRUE).  \code{...} are passed to
\code{filter_analysis_data}. See source code for available contrasts.
\code{test_data_FUN} is a function that takes combined dataset of an
element, distances, and the LOD and returns a dataset that can be
passed to a test statistic function.
}
\description{
Create a function that prepares data for analysis
}
