% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{create_annuli_idFUN}
\alias{create_annuli_idFUN}
\title{Create a function that identifies annual layers based on distance}
\usage{
create_annuli_idFUN(.data, .reference_transition = "on_")
}
\arguments{
\item{.data}{a measurement dataset (for a single id_transect)}

\item{.reference_transition}{the transition between layer to use as the reference.
Defaults to \code{"on_"} (the transition from "laser on" to the next layer).
\code{"ipx_"} is the transition from inner epoxy to the next layer (either nacre
or prismatic layer depending on the location of the transect)}
}
\value{
a function of \code{x} that \code{cut}s \code{x} based on the distances
measured between annuli
}
\description{
This creates a function that determines annuli layers:
\itemize{
\item only for transects that cross the nacre
\item only up to the ncr_psm transition
\item in the case that no annuli measurements exist the annual layer is labeled "U" for unknown
}
}
