% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{create_long_analysis_data}
\alias{create_long_analysis_data}
\title{Creates long form analytic dataset from the wide version}
\usage{
create_long_analysis_data(.wide_data)
}
\arguments{
\item{.wide_data}{a result of \code{\link{create_wide_analysis_data}}}
}
\description{
Creates long form analytic dataset from the wide version
}
