% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{create_qc_plot}
\alias{create_qc_plot}
\title{Create Plot for reviewing valve alignments}
\usage{
create_qc_plot(.idt_data, .edge_data, .ruler_length, .best_method)
}
\description{
Create Plot for reviewing valve alignments
}
