% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_data_functions.R
\name{create_transect_filter}
\alias{create_transect_filter}
\title{Create a function that filters a transect to selected observations}
\usage{
create_transect_filter(ds)
}
\arguments{
\item{ds}{the distance dataset which contains \code{distance}, \code{layer},
\code{annuli}, and \code{obs} variables}
}
\value{
a function that filters \code{ds} by chosen \code{.layers},
\code{annuli} (optionally), and (optionally trims the observations by
\code{.inner_buffer} and \code{.outer_buffer}.
}
\description{
Create a function that filters a transect to selected observations
}
