% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{create_wide_analysis_data}
\alias{create_wide_analysis_data}
\title{Creates a "wide" analytic dataset}
\usage{
create_wide_analysis_data(
  .valve_data,
  .reference_transition,
  .zero_function,
  .zf_args = list(),
  .include_ppm = FALSE
)
}
\arguments{
\item{.valve_data}{a \code{\link[tidyr]{nest}}ed dataset (one row per id_transect) with columns
\code{chemistry} (the chemistry data) and \code{measures} (the distance measurements)}

\item{.reference_transition}{passed to \code{\link{create_layer_idFUN}} and
\code{\link{create_annuli_idFUN}}}

\item{.zero_function}{passed to \code{\link{shift_distance}}}

\item{.zf_args}{passed to \code{\link{shift_distance}}}
}
\value{
an \code{\link[tidyr]{unnest}}ed \code{data.frame} containing:
\itemize{
\item id
\item transect
\item distance (shifted by \code{\link{shift_distance}})
\item layer: a label for the valve layer from which the observation is measured
\item annuli: a label for the annual layer from which the observation is measured
\item + one column for each element
} the \code{layer} and
\code{annuli} variables added
}
\description{
Creates a "wide" analytic dataset
}
