% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{munge_measurements}
\alias{munge_measurements}
\title{Create the data in "Bivalve Transect Datums}
\usage{
munge_measurements(.data)
}
\arguments{
\item{.data}{the result of \code{\link{read_measurements_sheet}}}
}
\description{
Create the data in "Bivalve Transect Datums
}
