% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_analysis_data.R
\name{ppm_to_mmol}
\alias{ppm_to_mmol}
\title{Convert ppm to mmol}
\usage{
ppm_to_mmol(ppm, gmol)
}
\arguments{
\item{ppm}{values in ppm}

\item{gmol}{mass of the element}
}
\description{
Convert ppm to mmol
}
