% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_data_functions.R
\name{ppm_to_mmol_camol}
\alias{ppm_to_mmol_camol}
\title{Convert ppm to mmol/Ca mol}
\usage{
ppm_to_mmol_camol(ppm, gmol, ca_wt_pct = 40.078, ca_ppm = 400432)
}
\arguments{
\item{ppm}{values in ppm}

\item{gmol}{mass of the element}
}
\description{
Convert ppm to mmol/Ca mol
}
