% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{read_measurements_sheet}
\alias{read_measurements_sheet}
\title{Import data in "Bivalve Transect Datums"}
\usage{
read_measurements_sheet(.inFile, .sheet)
}
\arguments{
\item{.inFile}{path to .xlsx file}

\item{.sheet}{which sheet to import}
}
\description{
Import data in "Bivalve Transect Datums"
}
