% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{shift_distance}
\alias{shift_distance}
\title{Shifts the distance variable in a chemistry dataset to a new 0 point based
on .zero_function}
\usage{
shift_distance(.chem_data, .zero_function = identity, .zf_args)
}
\arguments{
\item{.chem_data}{chemistry dataset for a single id_transect}

\item{.zero_function}{a function that shifts the distance function. Defaults to
\code{identity}. \code{x} must be first argument, where \code{x} is \code{.chem_data}}

\item{.zf_args}{arguments passed to \code{.zero_function}}
}
\description{
Shifts the distance variable in a chemistry dataset to a new 0 point based
on .zero_function
}
