% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{summarize_specimens}
\alias{summarize_specimens}
\title{Summarize specimens}
\usage{
summarize_specimens(adt, modify_layer = function(layer, annuli) {
    layer
})
}
\arguments{
\item{adt}{result from \code{analysis_data()}}

\item{modify_layer}{a function to modify the definition of the layer variable}
}
\description{
Summarize specimens
}
