% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_analysis_data.R
\name{filter_analysis_data}
\alias{filter_analysis_data}
\title{Create analysis data filter}
\usage{
filter_analysis_data(
  data,
  elements = "all",
  signals = "all",
  identifiers = rlang::exprs(),
  agrp = "all",
  transect_opts = list(.layers = "ncr")
)
}
\description{
Create analysis data filter
}
