% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_data_functions.R
\name{Z_maker}
\alias{Z_maker}
\title{Add a factor variable "Z" according to the specified constrast}
\usage{
Z_maker(contrast, dt)
}
\arguments{
\item{contrast}{name of the contrast}

\item{dt}{dataset with \code{site} variable}
}
\description{
Add a factor variable "Z" according to the specified constrast
}
